Backend - Queue Qr changes

+++ backend/app/Http/Controllers/Admin/QueuesController.php
@@ -3,13 +3,17 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Requests\Queues\CloseQueueRequest;
+use App\Http\Requests\Queues\QRExitRequest;
 use App\Models\Queue;
 use Illuminate\Http\Request;
 use App\Http\Requests\Queues\StoreQueueRequest;
 use App\Http\Requests\Queues\UpdateQueueRequest;
 use App\Models\File;
 use App\Models\Canteen;
+use App\Models\QueueStudent;
+use App\Models\Student;
 use Carbon\Carbon;
+use Illuminate\Support\Facades\Auth;
 
 class QueuesController extends Controller
 {
@@ -20,6 +24,15 @@ class QueuesController extends Controller
      */
     public function index(Canteen $canteen, Request $request) 
     {
+        $user = Auth::user();
+
+        if (!$user->canteens->contains('id', $canteen->id)) {
+            return response()->json([
+                'status' => 'error',
+                'permissions' => 'User does not have the right permissions to control this canteen.'
+            ], 403);
+        }
+
         $canteen->load([
             'school:id,name,file_id',
             'school.logo:id,current_name,path'
@@ -72,6 +85,15 @@ class QueuesController extends Controller
      */
     public function store(Canteen $canteen, StoreQueueRequest $request) 
     {
+        $user = Auth::user();
+
+        if (!$user->canteens->contains('id', $canteen->id)) {
+            return response()->json([
+                'status' => 'error',
+                'permissions' => 'User does not have the right permissions to control this canteen.'
+            ], 403);
+        }
+
         $today = Carbon::today();
 
         $activeQueues = Queue::where('canteen_id', $canteen->id)
@@ -112,6 +134,15 @@ class QueuesController extends Controller
      */
     public function show(Queue $queue) 
     {        
+        $user = Auth::user();
+
+        if (!$user->canteens->contains('id', $queue->canteen->id)) {
+            return response()->json([
+                'status' => 'error',
+                'permissions' => 'User does not have the right permissions to control this canteen.'
+            ], 403);
+        }
+
         return response()->json([
             'status' => 'success',
             'data' => [
@@ -130,6 +161,15 @@ class QueuesController extends Controller
      */
     public function update(Queue $queue, UpdateQueueRequest $request) 
     {
+        $user = Auth::user();
+
+        if (!$user->canteens->contains('id', $queue->canteen->id)) {
+            return response()->json([
+                'status' => 'error',
+                'permissions' => 'User does not have the right permissions to control this canteen.'
+            ], 403);
+        }
+
         $queue->update(array_merge(
             $request->validated(),
         ));
@@ -141,6 +181,15 @@ class QueuesController extends Controller
 
     public function close(Queue $queue, CloseQueueRequest $request) 
     {
+        $user = Auth::user();
+
+        if (!$user->canteens->contains('id', $queue->canteen->id)) {
+            return response()->json([
+                'status' => 'error',
+                'permissions' => 'User does not have the right permissions to control this canteen.'
+            ], 403);
+        }
+
         $queue->update(
             ['closed_at' => is_null($queue->closed_at) ? Carbon::now() : NULL]
         );
@@ -160,7 +209,38 @@ class QueuesController extends Controller
      */
     public function destroy(Queue $queue) 
     {
+        $user = Auth::user();
+
+        if (!$user->canteens->contains('id', $queue->canteen->id)) {
+            return response()->json([
+                'status' => 'error',
+                'permissions' => 'User does not have the right permissions to control this canteen.'
+            ], 403);
+        }
+
         $queue->delete();
+
+        return response()->json([
+            'status' => 'success'
+        ]);
+    }
+    
+    public function exit(Queue $queue, QRExitRequest $request){
+        $user = Auth::user();
+
+        if (!$user->canteens->contains('id', $queue->canteen->id)) {
+            return response()->json([
+                'status' => 'error',
+                'permissions' => 'User does not have the right permissions to control this canteen.'
+            ], 403);
+        }
+        
+        $queueStudent = QueueStudent::where([
+            'student_id' => $request->get('student_id'),
+            'queue_id' => $queue->id
+        ])->first();
+
+        $queueStudent->update(['exited_at' => Carbon::now()]);
 
         return response()->json([
             'status' => 'success'
